refactor: useChat hook
